Dashboard: add user full name and avatar.
* closes: inveniosoftware/invenio-app-rdm#1473
