+++ invenio_app_rdm/records_ui/views/__init__.py
@@ -15,13 +15,14 @@ from invenio_pidstore.errors import PIDDeletedError, PIDDoesNotExistError, \
 from invenio_records_resources.services.errors import PermissionDeniedError
 
 from ..searchapp import search_app_context
-from .deposits import dashboard, deposit_create, deposit_edit
+from .deposits import deposit_create, deposit_edit
 from .filters import can_list_files, get_scheme_label, has_previewable_files, \
     make_files_preview_compatible, order_entries, pid_url, \
     select_preview_file, to_previewer_files
-from .records import not_found_error, record_detail, record_export, \
-    record_file_download, record_file_preview, record_from_pid, \
-    record_latest, record_permission_denied_error, record_tombstone_error
+from .records import dashboard, not_found_error, record_detail, \
+    record_export, record_file_download, record_file_preview, \
+    record_from_pid, record_latest, record_permission_denied_error, \
+    record_tombstone_error
 
 
 #
